# Only optimize for size on NuttX