--- a/src/drivers/airspeed/CMakeLists.txt
+++ b/src/drivers/airspeed/CMakeLists.txt
@@ -33,10 +33,9 @@
 px4_add_module(
 	MODULE drivers__airspeed
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		airspeed.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :