--- a/src/drivers/blinkm/CMakeLists.txt
+++ b/src/drivers/blinkm/CMakeLists.txt
@@ -34,10 +34,9 @@ px4_add_module(
 	MODULE drivers__blinkm
 	MAIN blinkm
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		blinkm.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :