--- a/src/drivers/bma180/CMakeLists.txt
+++ b/src/drivers/bma180/CMakeLists.txt
@@ -34,10 +34,9 @@ px4_add_module(
 	MODULE drivers__bma180
 	MAIN bma180
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		bma180.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :