--- a/src/drivers/boards/asc-v1/CMakeLists.txt
+++ b/src/drivers/boards/asc-v1/CMakeLists.txt
@@ -34,7 +34,6 @@
 px4_add_module(
 	MODULE drivers__boards__asc-v1
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		../common/board_name.c
 		asc_init.c
@@ -45,4 +44,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :