--- a/src/drivers/boards/mindpx-v2/CMakeLists.txt
+++ b/src/drivers/boards/mindpx-v2/CMakeLists.txt
@@ -33,7 +33,6 @@
 px4_add_module(
 	MODULE drivers__boards__mindpx-v2
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		../common/board_dma_alloc.c
 		../common/board_name.c
@@ -46,4 +45,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :