--- a/src/drivers/boards/px4fmu-v1/CMakeLists.txt
+++ b/src/drivers/boards/px4fmu-v1/CMakeLists.txt
@@ -33,7 +33,6 @@
 px4_add_module(
 	MODULE drivers__boards__px4fmu-v1
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		../common/board_name.c
 		px4fmu_can.c
@@ -45,4 +44,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :