--- a/src/drivers/boards/sitl/CMakeLists.txt
+++ b/src/drivers/boards/sitl/CMakeLists.txt
@@ -33,10 +33,9 @@
 px4_add_module(
 	MODULE drivers__boards__sitl
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		sitl_led.c
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :