--- a/src/drivers/boards/tap-v1/CMakeLists.txt
+++ b/src/drivers/boards/tap-v1/CMakeLists.txt
@@ -34,7 +34,6 @@
 px4_add_module(
 	MODULE drivers__boards__tap-v1
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		../common/board_name.c
 		tap_init.c
@@ -48,4 +47,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :