--- a/src/drivers/bst/CMakeLists.txt
+++ b/src/drivers/bst/CMakeLists.txt
@@ -35,10 +35,9 @@ px4_add_module(
 	MAIN bst
 	STACK_MAIN 1200
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		bst.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :