--- a/src/drivers/camera_trigger/CMakeLists.txt
+++ b/src/drivers/camera_trigger/CMakeLists.txt
@@ -35,7 +35,6 @@ px4_add_module(
 	MAIN camera_trigger
 	STACK_MAIN 1200
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		camera_trigger.cpp
 		camera_trigger_params.c
@@ -45,4 +44,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :