--- a/src/drivers/ets_airspeed/CMakeLists.txt
+++ b/src/drivers/ets_airspeed/CMakeLists.txt
@@ -35,10 +35,9 @@ px4_add_module(
 	MAIN ets_airspeed
 	STACK_MAIN 1200
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		ets_airspeed.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :