--- a/src/drivers/hc_sr04/CMakeLists.txt
+++ b/src/drivers/hc_sr04/CMakeLists.txt
@@ -35,10 +35,9 @@ px4_add_module(
 	MAIN hc_sr04
 	STACK_MAIN 1200
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		hc_sr04.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :