--- a/src/drivers/irlock/CMakeLists.txt
+++ b/src/drivers/irlock/CMakeLists.txt
@@ -34,10 +34,9 @@ px4_add_module(
 	MODULE drivers__irlock
 	MAIN irlock
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		irlock.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :