--- a/src/drivers/l3gd20/CMakeLists.txt
+++ b/src/drivers/l3gd20/CMakeLists.txt
@@ -36,10 +36,9 @@ px4_add_module(
 	STACK_MAIN 1200
 	COMPILE_FLAGS
 		-Weffc++
-		-Os
 	SRCS
 		l3gd20.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :