--- a/src/drivers/led/CMakeLists.txt
+++ b/src/drivers/led/CMakeLists.txt
@@ -33,10 +33,9 @@
 px4_add_module(
 	MODULE drivers__led
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		led.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :