--- a/src/drivers/lis3mdl/CMakeLists.txt
+++ b/src/drivers/lis3mdl/CMakeLists.txt
@@ -36,7 +36,6 @@ px4_add_module(
 	STACK_MAIN 1200
 	COMPILE_FLAGS
 		-Weffc++
-		-Os
 	SRCS
 		lis3mdl_i2c.cpp
 		lis3mdl_spi.cpp
@@ -44,4 +43,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :