--- a/src/drivers/md25/CMakeLists.txt
+++ b/src/drivers/md25/CMakeLists.txt
@@ -34,11 +34,10 @@ px4_add_module(
 	MODULE drivers__md25
 	MAIN md25
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		md25.cpp
 		md25_main.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :