--- a/src/drivers/ms5611/CMakeLists.txt
+++ b/src/drivers/ms5611/CMakeLists.txt
@@ -41,9 +41,8 @@ px4_add_module(
 	MAIN ms5611
 	STACK_MAIN 1200
 	COMPILE_FLAGS
-		-Os
 	SRCS ${srcs}
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :