--- a/src/drivers/navio_gpio/CMakeLists.txt
+++ b/src/drivers/navio_gpio/CMakeLists.txt
@@ -34,7 +34,7 @@ px4_add_module(
 	MODULE drivers__navio_gpio
 	MAIN navio_gpio
 	STACK_MAIN 1200
-	COMPILE_FLAGS -Os
+	COMPILE_FLAGS
 	SRCS
 		navio_gpio.cpp
 	DEPENDS