--- a/src/drivers/navio_sysfs_rc_in/CMakeLists.txt
+++ b/src/drivers/navio_sysfs_rc_in/CMakeLists.txt
@@ -34,7 +34,7 @@ px4_add_module(
 	MODULE drivers__navio_sysfs_rc_in
 	MAIN navio_sysfs_rc_in
 	STACK_MAIN 1200
-	COMPILE_FLAGS -Os
+	COMPILE_FLAGS
 	SRCS
 		navio_sysfs_rc_in.cpp
 	DEPENDS