--- a/src/drivers/pwm_out_rc_in/CMakeLists.txt
+++ b/src/drivers/pwm_out_rc_in/CMakeLists.txt
@@ -34,7 +34,6 @@ px4_add_module(
 	MODULE drivers__pwm_out_rc_in
 	MAIN pwm_out_rc_in
 	COMPILE_FLAGS
-		-Os
 		-DMAVLINK_COMM_NUM_BUFFERS=1
 	SRCS
 		pwm_out_rc_in.cpp