--- a/src/drivers/px4io/CMakeLists.txt
+++ b/src/drivers/px4io/CMakeLists.txt
@@ -35,7 +35,6 @@ px4_add_module(
 	MAIN px4io
 	STACK_MAIN 1800
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		px4io.cpp
 		px4io_uploader.cpp
@@ -45,4 +44,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :