--- a/src/drivers/rgbled/CMakeLists.txt
+++ b/src/drivers/rgbled/CMakeLists.txt
@@ -34,10 +34,9 @@ px4_add_module(
 	MODULE drivers__rgbled
 	MAIN rgbled
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		rgbled.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :