--- a/src/drivers/roboclaw/CMakeLists.txt
+++ b/src/drivers/roboclaw/CMakeLists.txt
@@ -34,11 +34,10 @@ px4_add_module(
 	MODULE drivers__roboclaw
 	MAIN roboclaw
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		roboclaw_main.cpp
 		RoboClaw.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :