--- a/src/drivers/sf0x/CMakeLists.txt
+++ b/src/drivers/sf0x/CMakeLists.txt
@@ -34,11 +34,10 @@ px4_add_module(
 	MODULE drivers__sf0x
 	MAIN sf0x
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		sf0x.cpp
 		sf0x_parser.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :