--- a/src/drivers/sf1xx/CMakeLists.txt
+++ b/src/drivers/sf1xx/CMakeLists.txt
@@ -35,10 +35,9 @@ px4_add_module(
 	MODULE drivers__sf1xx
 	MAIN sf1xx
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		sf1xx.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :