--- a/src/drivers/srf02/CMakeLists.txt
+++ b/src/drivers/srf02/CMakeLists.txt
@@ -34,10 +34,9 @@ px4_add_module(
 	MODULE drivers__srf02
 	MAIN srf02
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		srf02.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :