--- a/src/drivers/srf02_i2c/CMakeLists.txt
+++ b/src/drivers/srf02_i2c/CMakeLists.txt
@@ -34,10 +34,9 @@ px4_add_module(
 	MODULE drivers__srf02_i2c
 	MAIN srf02_i2c
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		srf02_i2c.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :