--- a/src/drivers/stm32/adc/CMakeLists.txt
+++ b/src/drivers/stm32/adc/CMakeLists.txt
@@ -34,10 +34,9 @@ px4_add_module(
 	MODULE drivers__stm32__adc
 	MAIN adc
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		adc.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :