--- a/src/drivers/trone/CMakeLists.txt
+++ b/src/drivers/trone/CMakeLists.txt
@@ -35,10 +35,9 @@ px4_add_module(
 	MAIN trone
 	STACK_MAIN 1200
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		trone.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :