--- a/src/examples/ekf_att_pos_estimator/CMakeLists.txt
+++ b/src/examples/ekf_att_pos_estimator/CMakeLists.txt
@@ -34,7 +34,7 @@
 px4_add_module(
 	MODULE examples__ekf_att_pos_estimator
 	MAIN ekf_att_pos_estimator
-	COMPILE_FLAGS -Os
+	COMPILE_FLAGS
 	STACK_MAIN 3000
 	STACK_MAX 3400
 	SRCS
@@ -44,4 +44,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :