--- a/src/lib/controllib/CMakeLists.txt
+++ b/src/lib/controllib/CMakeLists.txt
@@ -33,7 +33,6 @@
 px4_add_module(
 	MODULE lib__controllib
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		blocks.cpp
 		block/Block.cpp
@@ -42,4 +41,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :