--- a/src/lib/conversion/CMakeLists.txt
+++ b/src/lib/conversion/CMakeLists.txt
@@ -33,10 +33,9 @@
 px4_add_module(
 	MODULE lib__conversion
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		rotation.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :