--- a/src/lib/external_lgpl/CMakeLists.txt
+++ b/src/lib/external_lgpl/CMakeLists.txt
@@ -33,10 +33,9 @@
 px4_add_module(
 	MODULE lib__external_lgpl
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		tecs/tecs.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :