--- a/src/lib/launchdetection/CMakeLists.txt
+++ b/src/lib/launchdetection/CMakeLists.txt
@@ -33,11 +33,10 @@
 px4_add_module(
 	MODULE lib__launchdetection
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		LaunchDetector.cpp
 		CatapultLaunchMethod.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :