--- a/src/lib/rc/CMakeLists.txt
+++ b/src/lib/rc/CMakeLists.txt
@@ -33,7 +33,6 @@
 px4_add_module(
 	MODULE lib__rc
 	COMPILE_FLAGS
-		-Os
 		-Wno-unused-result
 	SRCS
 		st24.c
@@ -43,4 +42,4 @@ px4_add_module(
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :