--- a/src/lib/runway_takeoff/CMakeLists.txt
+++ b/src/lib/runway_takeoff/CMakeLists.txt
@@ -33,10 +33,9 @@
 px4_add_module(
 	MODULE lib__runway_takeoff
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		RunwayTakeoff.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :