--- a/src/lib/terrain_estimation/CMakeLists.txt
+++ b/src/lib/terrain_estimation/CMakeLists.txt
@@ -34,10 +34,9 @@ px4_add_module(
 	MODULE lib__terrain_estimation
 	STACK_MAIN 1024
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		terrain_estimator.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :