--- a/src/modules/bottle_drop/CMakeLists.txt
+++ b/src/modules/bottle_drop/CMakeLists.txt
@@ -35,10 +35,9 @@ px4_add_module(
 	MAIN bottle_drop
 	STACK_MAIN 1200
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		bottle_drop.cpp
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :