--- a/src/modules/dataman/CMakeLists.txt
+++ b/src/modules/dataman/CMakeLists.txt
@@ -35,10 +35,9 @@ px4_add_module(
 	MAIN dataman
 	STACK_MAIN 1200
 	COMPILE_FLAGS
-		-Os
 	SRCS
 		dataman.c
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :