--- a/src/modules/ekf2/CMakeLists.txt
+++ b/src/modules/ekf2/CMakeLists.txt
@@ -33,7 +33,7 @@
 px4_add_module(
 	MODULE modules__ekf2
 	MAIN ekf2
-	COMPILE_FLAGS -Os
+	COMPILE_FLAGS
 	STACK_MAIN 2500
 	STACK_MAX 4000
 	SRCS
@@ -42,4 +42,4 @@ px4_add_module(
 		platforms__common
 		git_ecl
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :