--- a/src/modules/load_mon/CMakeLists.txt
+++ b/src/modules/load_mon/CMakeLists.txt
@@ -34,7 +34,7 @@ px4_add_module(
 	MODULE modules__load_mon
 	MAIN load_mon
 	STACK_MAIN 1200
-	COMPILE_FLAGS -Os
+	COMPILE_FLAGS
 	SRCS
 		load_mon.cpp
 	DEPENDS