--- a/src/modules/logger/CMakeLists.txt
+++ b/src/modules/logger/CMakeLists.txt
@@ -36,7 +36,7 @@ px4_add_module(
 	MAIN logger
 	PRIORITY "SCHED_PRIORITY_MAX-30"
 	STACK_MAIN 2200
-	COMPILE_FLAGS -Os
+	COMPILE_FLAGS
 	SRCS
 		logger.cpp
 		log_writer.cpp
@@ -44,4 +44,4 @@ px4_add_module(
 		platforms__common
 		modules__uORB
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :