--- a/src/modules/param/CMakeLists.txt
+++ b/src/modules/param/CMakeLists.txt
@@ -42,9 +42,8 @@ px4_generate_parameters_source(OUT param_files
 px4_add_module(
 	MODULE modules__param
 	COMPILE_FLAGS
-		-Os
 	SRCS ${param_files}
 	DEPENDS
 		platforms__common
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix : 
+# vim: set noet ft=cmake fenc=utf-8 ff=unix :