… src/modules/uORB/uORB_tests/CMakeLists.txt
… src/modules/uavcan/CMakeLists.txt
… src/modules/unit_test/CMakeLists.txt
… src/modules/vtol_att_control/CMakeLists.txt
… src/platforms/nuttx/CMakeLists.txt
… src/platforms/nuttx/px4_layer/CMakeLists.txt
… src/platforms/posix/drivers/adcsim/CMakeLists.txt
… src/platforms/posix/drivers/airspeedsim/CMakeLists.txt
… src/platforms/posix/drivers/barosim/CMakeLists.txt
… src/platforms/posix/drivers/gpssim/CMakeLists.txt
… src/platforms/posix/drivers/gyrosim/CMakeLists.txt
… src/platforms/posix/drivers/ledsim/CMakeLists.txt
… src/platforms/posix/drivers/rgbledsim/CMakeLists.txt
… src/platforms/posix/drivers/tonealrmsim/CMakeLists.txt
… src/platforms/posix/px4_layer/CMakeLists.txt
… src/platforms/posix/work_queue/CMakeLists.txt
… src/platforms/qurt/fc_addon/mpu_spi/CMakeLists.txt
… src/platforms/qurt/fc_addon/rc_receiver/CMakeLists.txt
… src/platforms/qurt/fc_addon/uart_esc/CMakeLists.txt
… src/platforms/qurt/px4_layer/CMakeLists.txt
… src/systemcmds/bl_update/CMakeLists.txt
… src/systemcmds/config/CMakeLists.txt
… src/systemcmds/dumpfile/CMakeLists.txt
… src/systemcmds/esc_calib/CMakeLists.txt
… src/systemcmds/i2c/CMakeLists.txt
… src/systemcmds/mixer/CMakeLists.txt
… src/systemcmds/motor_test/CMakeLists.txt
… src/systemcmds/mtd/CMakeLists.txt
… src/systemcmds/nshterm/CMakeLists.txt
… src/systemcmds/param/CMakeLists.txt