With this change only builds for NuttX which are very much flash size
constraint are optimized for size. All other builds (e.g. SITL,
Snapdragon, etc.) are left at the default for debugging or -O2 for the
usual use.

This addresses #4759.

## Build times
### POSIX:

-Os:
`make posix 164.70s user 11.82s system 334% cpu 52.833 total`

-O2:
`make posix 187.23s user 12.61s system 351% cpu 56.923 total`

(default no -O2, not in this PR):
`make posix 115.52s user 11.32s system 337% cpu 37.601 total`

### QURT:

with -Os
`make qurt_eagle_default  257.10s user 6.22s system 208% cpu 2:06.03 total`

-O2
`make qurt_eagle_default  245.09s user 5.87s system 221% cpu 1:53.15 total`

(default no -O2, not in this PR):
`make qurt_eagle_default  98.16s user 5.94s system 341% cpu 30.496 total`


## size comparison:

### px4fmu-v2_default

before (-Os):
`1018992 build_px4fmu-v2_default/src/firmware/nuttx/firmware_nuttx.bin`

after (-Os):
`1018992 build_px4fmu-v2_default/src/firmware/nuttx/firmware_nuttx.bin`

Unfortunately, the changes here include the latest ecl submodule, so this can't be merged before the rest of ecl has been merged.