@catch-twenty-two 2016-09-27
> before (-Os):
> 1018992 build_px4fmu-v2_default/src/firmware/nuttx/firmware_nuttx.bin
> 
> after (-Os):
> 1018992 build_px4fmu-v2_default/src/firmware/nuttx/firmware_nuttx.bin

Just curious, are these supposed to show a difference?