@dagar 2016-09-27
This is definitely than specifying optimization level in every module. A step further would probably be utilizing the existing cmake build types.

      CMAKE_CXX_FLAGS
      CMAKE_CXX_FLAGS_DEBUG           -g
      CMAKE_CXX_FLAGS_MINSIZEREL      -Os -DNDEBUG
      CMAKE_CXX_FLAGS_RELEASE         -O3 -DNDEBUG
      CMAKE_CXX_FLAGS_RELWITHDEBINFO  -O2 -g