@julianoes 2016-09-28
@catch-twenty-two No, they show that the size has not changed :) However, the size of this PR is currently bigger than master but that's because the ecl submodule is updated.

@dagar there is quite a bit of stuff in cmake about the optimization flags, so I'm not sure how to change it to the cmake flags you mentionened.