@bkueng 2016-09-30
rebased & applied