@julianoes 2016-09-30
> Unfortunately, the changes here include the latest ecl submodule, so this can't be merged before the rest of ecl has been merged.

Uh oh. Have all the ecl changes gone in?